Changed: begin refactor
