+++ model/event.go
@@ -1,0 +1,7 @@
+package model
+
+type Event struct {
+	Volume int     `json:"volume"`
+	Music  []Music `json:"music"`
+	Sfx    []Sfx   `json:"sfx"`
+}
