+++ model/model.go
@@ -1,0 +1,5 @@
+package model
+
+type LocalFile interface {
+	func GetPath() string
+}
