+++ model/sate.go
@@ -1,0 +1,5 @@
+package model
+
+type State struct {
+	States []string `json:"states"`
+}
