+++ model/sfx.go
@@ -1,0 +1,4 @@
+package model
+
+type Sfx struct {
+}
